WIP: NetFileFuzzer moved to main snapchage
